# access-denied-backend/docs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "access-denied",
        "description": "API Documentation",
        "version": "unknown"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8083/access-denied-web-api",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/roles/{rolesName}": {
            "get": {
                "tags": ["Role Management"],
                "operationId": "getRole",
                "parameters": [
                    {
                        "name": "rolesName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AccessDenied-API-Key": []
                    }
                ]
            },
            "put": {
                "tags": ["Role Management"],
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "rolesName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRoleRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AccessDenied-API-Key": []
                    }
                ]
            },
            "delete": {
                "tags": ["Role Management"],
                "operationId": "deleteRole",
                "parameters": [
                    {
                        "name": "rolesName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "AccessDenied-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AccessDeniedNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "accessdenied"
                    }
                }
            ]
        },
        "/api/v1/privileges/{privilegeName}": {
            "get": {
                "tags": ["Privilege Management"],
                "operationId": "getPrivilege",
                "parameters": [
                    {
                        "name": "privilegeName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Privilege"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AccessDenied-API-Key": []
                    }
                ]
            },
            "put": {
                "tags": ["Privilege Management"],
                "operationId": "updatePrivilege",
                "parameters": [
                    {
                        "name": "privilegeName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePrivilegeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Privilege"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AccessDenied-API-Key": []
                    }
                ]
            },
            "delete": {
                "tags": ["Privilege Management"],
                "operationId": "deletePrivilege",
                "parameters": [
                    {
                        "name": "privilegeName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "AccessDenied-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AccessDeniedNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "accessdenied"
                    }
                }
            ]
        },
...
        "/api/v1/roles/": {
            "get": {
                "tags": ["Role Management"],
                "operationId": "listRoles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AccessDenied-API-Key": []
                    }
                ]
            },
            "post": {
                "tags": ["Role Management"],
                "operationId": "createRole",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRoleRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AccessDenied-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AccessDeniedNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "accessdenied"
                    }
                }
            ]
        },
        "/api/v1/privileges/": {
            "get": {
                "tags": ["Privilege Management"],
                "operationId": "listPrivileges",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AccessDenied-API-Key": []
                    }
                ]
            },
            "post": {
                "tags": ["Privilege Management"],
                "operationId": "createPrivilege",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePrivilegeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Privilege"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AccessDenied-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AccessDeniedNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "accessdenied"
                    }
                }
            ]
        },
...
    },
    "components": {
        "schemas": {
...
            "Privilege": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "Role": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "privileges": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Privilege"
                        }
                    }
                }
            },
...
            "UpdateRoleRequest": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "privileges": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "UpdatePrivilegeRequest": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    }
                }
            },
...
            "CreateRoleRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "privileges": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "CreatePrivilegeRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "AccessDenied-API-Key": {
                "type": "apiKey",
                "name": "AccessDenied-API-Key",
                "in": "header"
            }
        }
    }
}